Separated parsing functions into a separate module
